conv 1 channel

+++ Algorithms/ADNN/activations.jl
@@ -1,4 +1,4 @@
 
-function sigma(x::Float64)
+function sigma(x)
     1/(1+exp(-x))
 end
